Fixed compiler warnings.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@11360 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/states_screens/cutscene_gui.cpp
@@ -41,10 +41,11 @@ void CutsceneGUI::renderGlobal(float dt)
 {
     if (m_fade_level > 0.0f)
     {
-        irr_driver->getVideoDriver()->draw2DRectangle(video::SColor(m_fade_level*255, 0,0,0),
-                                                      core::rect<s32>(0, 0,
-                                                                      UserConfigParams::m_width,
-                                                                      UserConfigParams::m_height));
+        irr_driver->getVideoDriver()->draw2DRectangle(
+                                video::SColor((int)(m_fade_level*255), 0,0,0),
+                                core::rect<s32>(0, 0,
+                                                UserConfigParams::m_width,
+                                                UserConfigParams::m_height));
     }
     
     if (m_subtitle.size() > 0)
